+++ airbyte-webapp/src/pages/AuthPage/components/FormHeaderSection.tsx
@@ -1,8 +1,9 @@
-import React from "react"; // , { useState }
+import React, { useState } from "react";
 
 import { Link } from "components";
-// import { GlobIcon } from "components/icons/GlobIcon";
-// import { LOCALES } from "locales";
+import { GlobIcon } from "components/icons/GlobIcon";
+
+import { LOCALES } from "locales";
 
 import styles from "./FormHeaderSection.module.scss";
 
@@ -13,12 +14,12 @@ interface Iporps {
 }
 
 export const FormHeaderSection: React.FC<Iporps> = ({ text, link, buttonText }) => {
-  // const [language] = useState<string>(LOCALES.ENGLISH); // setLanguage
+  const [language] = useState<string>(LOCALES.ENGLISH); // setLanguage
 
   return (
     <div className={styles.head}>
       <div className={styles.selectBox}>
-        {/* <div className={styles.globalIcon}>
+        <div className={styles.globalIcon}>
           <GlobIcon color="#374151" />
         </div>
         <select
@@ -33,7 +34,7 @@ export const FormHeaderSection: React.FC<Iporps> = ({ text, link, buttonText }) 
         >
           <option value={LOCALES.ENGLISH}>English</option>
           <option value={LOCALES.CHINESE_SIMPLIFIED}>简体中文</option>
-        </select> */}
+        </select>
       </div>
       <div className={styles.headRight}>
         <div className={styles.headRightText}>{text}</div>
